Convert inline methods with immediates to macros
The used NEON intrinsics have an immediate parameter which has to
be known at compile time because it's part of the instruction.

Unlike GCC, clang is not able figure this out during inlining.
Because of that, affected inline methods have been converted to macros.

Also removed zero-test for rshiftbyte_m128 and lshiftbyte_m128 since
it's not relevant to the hyperscan implementation. The expansion of
the macro for lshiftbyte_m128(a, 0) in the simd_utils tests caused
a compile error due to a failing range check. The underlying intrinsic
only accepts values between 0 and 15. But passing in zero causes the
macro to expand the value to 16 - 0, which makes clang fail. The explicit
check for zero doesn't make a difference, the error still get's thrown.
Since it's not relevant to the actual implementation, i've decided to
remove it.